Work on pre-integrated specular

+++ core/rendering/texture/sampler2d.go
@@ -6,5 +6,5 @@ import (
 
 type Sampler2D interface {
 	Sample(uv math.Vector2) math.Vector4
-	SampleLod(uv math.Vector2, mipLevel int) math.Vector4
+	SampleLod(uv math.Vector2, mipLevel float32) math.Vector4
 }
