+++ core/rendering/texture/sampler2d_linear.go
@@ -47,10 +47,10 @@ func (sampler *Sampler2D_linear) Sample(uv math.Vector2) math.Vector4 {
 	return bilinear(c00, c01, c10, c11, s, t)
 }
 
-func (s *Sampler2D_linear) SampleLod(uv math.Vector2, mipLevel int) math.Vector4 {
+func (s *Sampler2D_linear) SampleLod(uv math.Vector2, mipLevel float32) math.Vector4 {
 	auv := s.address.address2D(uv)
 
-	b := &s.texture.Image.Buffers[mipLevel]
+	b := &s.texture.Image.Buffers[int(mipLevel)]
 
 	x := int(auv.X * float32(b.dimensions.X - 1) + 0.5)
 	y := int(auv.Y * float32(b.dimensions.Y - 1) + 0.5)
